Rename library to "SNMP_Agent" (#24)
* Setup Library for inclusion in Arduino Library Manager

* Change Library name to SNMP_Agent
