+++ .github/workflows/test.yml
@@ -13,6 +13,8 @@ jobs:
 
     steps:
     - uses: actions/checkout@v2
+    - name: arduino/arduino-lint-action
+      uses: arduino/arduino-lint-action@v1.0.0
     - name: run tests
       run: |
         cd tests
