+++ library.properties
@@ -1,9 +1,10 @@
-name=Arduino_SNMP2
+name=SNMP_Agent
 version=2.0.0
 author=Aidan Cyr <cyraidan@gmail.com>
 maintainer=Aidan Cyr <cyraidan@gmail.com>
-sentence=An SNMPv2c Agent for Arduino
-paragraph=
+sentence=Arduino_SNMP: An SNMPv2c Agent for Arduino
+paragraph=(Previously Arduino_SNMP)
 category=Communication
 architectures=esp32
-includes=Arduino_SNMP.h, SNMPTrap.h
+includes=SNMP_Agent.h, SNMPTrap.h
+url=https://github.com/0neblock/Arduino_SNMP
